pyproject.toml
[tool.poetry]
name = "copa-ata"
version = "0.1.0"
description = "Automatic ATA assignment to event"
authors = ["anand-kamble <amk23j@fsu.edu>", "erlebach <gerlebacher@fsu.edu>"]

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.2.0"
langchain-core = "^0.2.0"
langchain-community = "^0.2.0"
langsmith = "^0.1.59"
openai = "^1.30.1"
python-dotenv = "^1.0.1"
huggingface-hub = "^0.23.0"
langchain-huggingface = "^0.0.1"
sentencepiece = "^0.2.0"
protobuf = "^5.27.0"
pandas = "^2.2.2"
torch = "^2.3.0"

[tool.poetry.scripts]
ata="copa-ata.main:main"


[tool.poetry.group.dev.dependencies]
jupyter = "^1.0.0"
ipython = "^8.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
